fix(tooltips): like/select popups overlap
Fixes #31

This commit hides the `select ...` tooltip message
when the ShowSmall component only *does not disable*
liking, i.e. (currently) when not showing select

these are the current 2 variants:
1. Select
2. Like

But if this ever changes this might be another source of bugs. Probably having an enum for all the variants
would make more sense, but w/e.